# routes.js
const express=require('express'),jwt=require('jsonwebtoken'),bcrypt=require('bcryptjs'),crypto=require('crypto');const {runScraper}=require('./scraper');
const clean=(v,max=5000)=>String(v??'').replace(/[<>\u0000-\u001f]/g,'').trim().slice(0,max);const json=v=>{try{return JSON.parse(v||'[]')}catch{return[]}};
module.exports=function routes(db,{loginLimiter,tipLimiter,scrapeLimiter}){const r=express.Router();const secret=process.env.JWT_SECRET||crypto.createHash('sha256').update('development-only-'+process.cwd()).digest('hex');
function auth(req,res,next){try{req.user=jwt.verify(req.cookies.admin_token,secret);next()}catch{return res.status(401).json({error:{code:'UNAUTHORIZED',message:'Authentication required'}})}}function csrf(req,res,next){if(!['GET','HEAD'].includes(req.method)&&(!req.headers['x-csrf-token']||req.headers['x-csrf-token']!==req.cookies.csrf_token))return res.status(403).json({error:{code:'CSRF',message:'Invalid CSRF token'}});next()}
r.get('/timer',(q,s)=>s.json(db.prepare('SELECT * FROM timer_configuration WHERE id=1').get()));
r.get('/posts',(q,s)=>{const page=Math.max(1,+q.query.page||1),limit=Math.min(50,Math.max(1,+q.query.limit||10)),search=`%${clean(q.query.search||'',100)}%`,cat=clean(q.query.category||'',100);const where=`status='published' AND (title LIKE ? OR summary LIKE ?) ${cat?'AND category=?':''}`;const args=cat?[search,search,cat]:[search,search];const total=db.prepare(`SELECT count(*) n FROM blog_posts WHERE ${where}`).get(...args).n;const items=db.prepare(`SELECT * FROM blog_posts WHERE ${where} ORDER BY published_at DESC LIMIT ? OFFSET ?`).all(...args,limit,(page-1)*limit).map(p=>({...p,tags:json(p.tags),related_entities:json(p.related_entities)}));s.json({items,page,total,pages:Math.ceil(total/limit)})});
r.get('/posts/:id',(q,s)=>{const p=db.prepare(`SELECT * FROM blog_posts WHERE status='published' AND (id=? OR slug=?)`).get(q.params.id,q.params.id);if(!p)return s.status(404).json({error:{code:'NOT_FOUND',message:'Post not found'}});s.json({...p,tags:json(p.tags),related_entities:json(p.related_entities)})});
for(const [path,table] of [['milestones','milestones'],['timeline','timeline_events'],['resources','resources'],['faqs','faqs']])r.get('/'+path,(q,s)=>s.json(db.prepare(`SELECT * FROM ${table} ORDER BY ${table==='faqs'?'sort_order':'id'} DESC`).all()));
r.get('/dashboard',(q,s)=>s.json({pending:db.prepare("SELECT count(*) n FROM blog_posts WHERE status='pending_review'").get().n,published:db.prepare("SELECT count(*) n FROM blog_posts WHERE status='published'").get().n,quote:db.prepare('SELECT * FROM quotes WHERE active=1 ORDER BY id LIMIT 1').get()}));
r.post('/tips',tipLimiter,(q,s)=>{const m=clean(q.body.message,2000);if(m.length<10)return s.status(400).json({error:{code:'VALIDATION',message:'Message must be at least 10 characters'}});db.prepare('INSERT INTO tips(name,email,message) VALUES(?,?,?)').run(clean(q.body.name,100),clean(q.body.email,200),m);s.status(201).json({message:'Tip received for moderation.'})});
r.post('/analytics',(q,s)=>{db.prepare('INSERT INTO analytics_events(event_type,path) VALUES(?,?)').run(clean(q.body.event_type||'page_view',50),clean(q.body.path,200));s.status(202).end()});
r.post('/auth/login',loginLimiter,async(q,s)=>{const u=db.prepare('SELECT * FROM admin_users WHERE email=?').get(clean(q.body.email,200).toLowerCase());if(!u||!await bcrypt.compare(String(q.body.password||''),u.password_hash))return s.status(401).json({error:{code:'INVALID_LOGIN',message:'Invalid email or password'}});const token=jwt.sign({sub:u.id,email:u.email},secret,{expiresIn:'30m'}),c=crypto.randomBytes(24).toString('hex'),secure=process.env.NODE_ENV==='production';s.cookie('admin_token',token,{httpOnly:true,sameSite:'strict',secure,maxAge:1800000}).cookie('csrf_token',c,{sameSite:'strict',secure,maxAge:1800000}).json({csrfToken:c,mustChangePassword:!!u.must_change_password})});
r.post('/auth/logout',auth,csrf,(q,s)=>s.clearCookie('admin_token').clearCookie('csrf_token').status(204).end());r.get('/admin/session',auth,(q,s)=>s.json({user:q.user,csrfToken:q.cookies.csrf_token}));
r.get('/admin/posts',auth,(q,s)=>s.json(db.prepare('SELECT * FROM blog_posts ORDER BY created_at DESC').all()));
r.patch('/admin/posts/:id',auth,csrf,(q,s)=>{const allowed=['title','summary','content','category','status','source_name','source_url','tags','related_entities'];const values={},params=[];for(const k of allowed)if(k in q.body){values[k]=k==='tags'||k==='related_entities'?JSON.stringify(q.body[k]):clean(q.body[k]);params.push(`${k}=@${k}`)}if(!params.length)return s.status(400).json({error:{code:'VALIDATION',message:'No editable fields'}});values.id=q.params.id;db.prepare(`UPDATE blog_posts SET ${params.join(',')},updated_at=CURRENT_TIMESTAMP,published_at=CASE WHEN @status='published' THEN COALESCE(published_at,CURRENT_TIMESTAMP) ELSE published_at END WHERE id=@id`).run({status:q.body.status||'',...values});s.json({ok:true})});
r.put('/admin/timer',auth,csrf,(q,s)=>{const status=clean(q.body.status,40),impl=clean(q.body.implementation_date,100)||null;if(!['monitoring','under_discussion','officially_implemented'].includes(status)||status==='officially_implemented'&&(!impl||!clean(q.body.official_source_url,1000)))return s.status(400).json({error:{code:'VALIDATION',message:'Official implementation requires date and source URL'}});db.prepare('UPDATE timer_configuration SET status=?,is_running=?,implementation_date=?,decision_note=?,official_source_url=?,updated_at=CURRENT_TIMESTAMP WHERE id=1').run(status,status==='officially_implemented'?0:1,impl,clean(q.body.decision_note),clean(q.body.official_source_url,1000));s.json(db.prepare('SELECT * FROM timer_configuration WHERE id=1').get())});
r.post('/admin/scraper/run',auth,csrf,scrapeLimiter,async(q,s,next)=>{try{s.json(await runScraper(db,q.body.mock?{fixtures:q.body.fixtures||[]}:{}))}catch(e){next(e)}});r.get('/admin/logs',auth,(q,s)=>s.json({runs:db.prepare('SELECT * FROM scraper_runs ORDER BY id DESC LIMIT 100').all(),logs:db.prepare('SELECT * FROM system_logs ORDER BY id DESC LIMIT 100').all()}));r.get('/admin/analytics',auth,(q,s)=>s.json(db.prepare('SELECT event_type,count(*) count FROM analytics_events GROUP BY event_type').all()));r.get('/admin/tips',auth,(q,s)=>s.json(db.prepare('SELECT * FROM tips ORDER BY id DESC').all()));return r};